+++ .eslintrc.cjs
@@ -4,7 +4,7 @@ module.exports = {
   parser: '@typescript-eslint/parser',
   parserOptions: { ecmaVersion: 'latest', sourceType: 'module', ecmaFeatures: { jsx: true } },
   plugins: ['@typescript-eslint', 'react-hooks'],
-  ignorePatterns: ['dist/', 'dist-electron/', 'release/', 'android/', 'node_modules/'],
+  ignorePatterns: ['dist/', 'dist-electron/', 'release/', 'android/', 'node_modules/', 'sdk/dist/', 'sdk/build-src/', 'sdk/sample-app/'],
   rules: {
     'react-hooks/rules-of-hooks': 'error'
   }